bugfixes and improvements
- Fix Sample::convertStereoToMono for 8-bit stereo WAVs.
- Speed up Sample::reverse.
- Speed up loading 8-bit WAVs.
- Fix race condition in sample preview logic.

+++ libntxm/include/ntxm/fifocommand.h
@@ -45,7 +45,7 @@ struct PlaySampleCommand
 /* Command parameters for stopping a sample */
 struct StopSampleSoundCommand
 {
-    int channel;
+    u8 channel;
 };
 
 /* Command parameters for starting to record from the microphone */
